Add all used files to the cleaner.
This patch touches #44.
This patch touches @dnmTX

Indeed, before this patch we were not cleaning all files we use which are
located in the `output/` directory. This caused some noises.

In fact, even if we cleaned what we distributed (clean, domains, volatile,
whitelisted), a removed subject could have been in the backend.
Causing it to be reintroduced.

Service Notice (cc @Ultimate-Hosts-Blacklist/masters)
  * Simply remove what we distribute (except `domains.list`)
    * Literally clean, volatile, ips, and whitelisted.
  * Then, restart from the beginning with a commit which contain
    `Launch test` (case sensitive) in its message.